用线程阻塞出队处理玩家匹配的功能。

+++ src/main/java/com/artemis/football/model/MatchRoom.java
@@ -34,12 +34,12 @@ public class MatchRoom {
     /**
      * 角度
      */
-    private Integer angle;
+    private Float angle;
 
     /**
      * 力度
      */
-    private Integer power;
+    private Float power;
 
     @JsonIgnore
     private Lock lock = new ReentrantLock();
